Extract AdminUserController#login_as
Moves `AdminUserController#login_as` to its own controller.
