+++ app/controllers/admin_user_controller.rb
@@ -12,7 +12,6 @@ class AdminUserController < AdminController
                                             :update,
                                             :show_bounce_message,
                                             :clear_bounce,
-                                            :login_as,
                                             :clear_profile_photo ]
 
   before_filter :clear_roles,
@@ -98,21 +97,6 @@ class AdminUserController < AdminController
     @admin_user.email_bounce_message = ""
     @admin_user.save!
     redirect_to admin_user_url(@admin_user)
-  end
-
-  def login_as
-    if cannot? :login_as, @admin_user
-      flash[:error] =
-        "You don't have permission to log in as #{ @admin_user.name }"
-      return redirect_to admin_user_path(@admin_user)
-    end
-
-    @admin_user.confirm!
-
-    session[:user_id] = @admin_user.id
-    session[:user_circumstance] = 'login_as'
-
-    redirect_to user_path(@admin_user)
   end
 
   def clear_profile_photo
